feat: add Docker action
Add a workflow that takes a Dockerfile, target, a registry, and desired
platforms, builds images, and pushes them to the registry.

Keeps some of the inputs from the Goreleaser workflow, but not all. I
removed some stuff that looked unecessary (because it could be done in
the Dockerfile instead) or where it looked like there were better
options for reasonable defaults.

Signed-off-by: Travis Raines <571832+rainest@users.noreply.github.com>